Refactor break screen panel
