+++ src/dashboard/breakScreen/main.ts
@@ -1,0 +1,7 @@
+import '../globalStyles.css';
+import './breakScreen.css';
+
+import './mainScene';
+import './teams';
+import './rounds';
+import './scenes';
